Shift functionality into scmdataframe.base
