+++ openscm/scmdataframe/__init__.py
@@ -3,8 +3,6 @@ The OpenSCM high-level API provides high-level functionality around
 single model runs.  This includes reading/writing input and output
 data, easy setting of parameters and stochastic ensemble runs.
 """
-import datetime
-
 import numpy as np
 import pandas as pd
 from dateutil import parser
@@ -14,12 +12,7 @@ from openscm.parameters import ParameterType
 from openscm.units import UnitConverter
 from openscm.utils import convert_openscm_time_to_datetime
 
-from .base import ScmDataFrameBase
-
-try:
-    from pyam import IamDataFrame
-except ImportError:
-    IamDataFrame = None
+from .base import ScmDataFrameBase, IamDataFrame, df_append
 
 ONE_YEAR_IN_S_INTEGER = int(round(UnitConverter("year", "s").convert_from(1)))
 
@@ -35,96 +28,6 @@ class ScmDataFrame(ScmDataFrameBase):
     your data frame to an IamDataFrame first with `to_iamdataframe()`. Note: this
     operation can be relatively computationally expensive for large data sets.
     """
-
-    def append(self, other: ScmDataFrameBase, inplace=False, **kwargs):
-        """Appends additional timeseries from a castable object to the current dataframe
-
-        See ``df_append``
-
-        Parameters
-        ----------
-        other: openscm.scmdataframe.ScmDataFrame or something which can be cast to ScmDataFrameBase
-        """
-        if not isinstance(other, ScmDataFrameBase):
-            other = self.__class__(other, **kwargs)
-
-        return df_append([self, other], inplace=inplace)
-
-    def to_iamdataframe(self):
-        """Convert to  IamDataFrame instance
-
-        Returns
-        -------
-        An pyam.IamDataFrame instance containing the same data
-        """
-        if IamDataFrame is None:
-            raise ImportError(
-                "pyam is not installed. Features involving IamDataFrame are unavailable"
-            )
-
-        class LongIamDataFrame(IamDataFrame):
-            """This baseclass is a custom implementation of the IamDataFrame which handles datetime data which spans longer than pd.to_datetime
-            can handle
-            """
-
-            def _format_datetime_col(self):
-                if isinstance(self.data["time"].iloc[0], str):
-
-                    def convert_str_to_datetime(inp):
-                        return parser.parse(inp)
-
-                    self.data["time"] = self.data["time"].apply(convert_str_to_datetime)
-
-                not_datetime = [
-                    not isinstance(x, datetime.datetime) for x in self.data["time"]
-                ]
-                if any(not_datetime):
-                    bad_values = self.data[not_datetime]["time"]
-                    error_msg = "All time values must be convertible to datetime. The following values are not:\n{}".format(
-                        bad_values
-                    )
-                    raise ValueError(error_msg)
-
-        return LongIamDataFrame(self.timeseries())
-
-    def to_csv(self, path: str, **kwargs):
-        """Write timeseries data to a csv file
-
-        Parameters
-        ----------
-        path: string
-            file path
-        """
-        self.to_iamdataframe().to_csv(path, **kwargs)
-
-    def line_plot(self, x: str = "time", y: str = "value", **kwargs):
-        """Helper to generate line plots of timeseries
-
-        See ``pyam.IamDataFrame.line_plot`` for more information
-
-        """
-        return self.to_iamdataframe().line_plot(x, y, **kwargs)
-
-    def scatter(self, x: str, y: str, **kwargs):
-        """Plot a scatter chart using metadata columns
-
-        see pyam.plotting.scatter() for all available options
-        """
-        self.to_iamdataframe().scatter(x, y, **kwargs)
-
-    def region_plot(self, **kwargs):
-        """Plot regional data for a single model, scenario, variable, and year
-
-        see ``pyam.plotting.region_plot()`` for all available options
-        """
-        return self.to_iamdataframe().region_plot(**kwargs)
-
-    def pivot_table(self, index, columns, **kwargs):
-        """Returns a pivot table
-
-        see ``pyam.core.IamDataFrame.pivot_table()`` for all available options
-        """
-        return self.to_iamdataframe().pivot_table(index, columns, **kwargs)
 
 
 def convert_core_to_scmdataframe(
@@ -230,71 +133,3 @@ def convert_config_dict_to_parameter_set(config):
         view.set(value.magnitude)
 
     return parameters
-
-
-def df_append(dfs, inplace=False):
-    """
-    Append together many dataframes into a single ScmDataFrame
-    When appending many dataframes it may be more efficient to call this routine once with a list of ScmDataFrames, then using
-    `ScmDataFrame.append`. If timeseries with duplicate metadata are found, the timeseries are appended. For duplicate timeseries,
-    values fallings on the same timestep are averaged.
-    Parameters
-    ----------
-    dfs: list of ScmDataFrameBase object, string or pd.DataFrame.
-    The dataframes to append. Values will be attempted to be cast to non ScmDataFrameBase.
-    inplace : bool
-    If True, then the operation updates the first item in dfs
-    Returns
-    -------
-    ScmDataFrameBase-like object containing the merged data. The resultant class will be determined by the type of the first object
-    in dfs
-    """
-    dfs = [
-        df if isinstance(df, ScmDataFrameBase) else ScmDataFrameBase(df) for df in dfs
-    ]
-    joint_dfs = [df.copy() for df in dfs]
-    joint_meta = []
-    for df in joint_dfs:
-        joint_meta += df.meta.columns.tolist()
-
-    joint_meta = set(joint_meta)
-
-    # should probably solve this https://github.com/pandas-dev/pandas/issues/3729
-    na_fill_value = -999
-    for i, _ in enumerate(joint_dfs):
-        for col in joint_meta:
-            if col not in joint_dfs[i].meta:
-                joint_dfs[i].set_meta(na_fill_value, name=col)
-
-    # we want to put data into timeseries format and pass into format_ts instead of format_data
-    data = pd.concat(
-        [d.timeseries().reorder_levels(joint_meta) for d in joint_dfs], sort=False
-    )
-
-    data = data.reset_index()
-    data[list(joint_meta)] = data[joint_meta].replace(
-        to_replace=np.nan, value=na_fill_value
-    )
-    data = data.set_index(list(joint_meta))
-
-    data = data.groupby(data.index.names).mean()
-
-    if not inplace:
-        ret = dfs[0].copy()
-    else:
-        ret = dfs[0]
-
-    ret._data = data.reset_index(drop=True).T
-    ret._data = ret._data.sort_index()
-    ret["time"] = ret._data.index.values
-    ret._data = ret._data.astype(float)
-
-    ret._meta = (
-        data.index.to_frame()
-        .reset_index(drop=True)
-        .replace(to_replace=na_fill_value, value=np.nan)
-    )
-    ret._sort_meta_cols()
-
-    if not inplace:
-        return ret
