+++ openscm/scmdataframe/base.py
@@ -2,12 +2,17 @@ from __future__ import annotations
 
 import copy
 import os
-from datetime import datetime
+import datetime
 from logging import getLogger
+from typing import Dict, Union
 
 import numpy as np
 import pandas as pd
 from dateutil import parser
+try:
+    from pyam import IamDataFrame
+except ImportError:
+    IamDataFrame = None
 
 from openscm.core import Core
 from openscm.timeseries_converter import (
@@ -107,13 +112,13 @@ def format_data(df):
         cols = set(df.columns) - set(REQUIRED_COLS)
         time_cols, extra_cols = False, []
         for i in cols:
-            if is_floatlike(i) or isinstance(i, datetime):
+            if is_floatlike(i) or isinstance(i, datetime.datetime):
                 time_cols = True
             else:
                 try:
                     try:
                         # most common format
-                        datetime.strptime(i, "%Y-%m-%d %H:%M:%S")
+                        datetime.datetime.strptime(i, "%Y-%m-%d %H:%M:%S")
                         time_cols = True
                     except ValueError:
                         # this is super slow so avoid if possible
@@ -185,7 +190,11 @@ class ScmDataFrameBase(object):
     """
 
     def __init__(
-        self, data, columns=None, climate_model: str = "unspecified", **kwargs
+        self,
+        data: Union[IamDataFrame, pd.DataFrame, np.ndarray, str],
+        columns: Dict[str, list] = None,
+        climate_model: str = "unspecified",
+        **kwargs
     ):
         """Initialize an instance of an ScmDataFrameBase
 
@@ -234,7 +243,7 @@ class ScmDataFrameBase(object):
         elif isinstance(data, pd.DataFrame) or isinstance(data, pd.Series):
             (_df, _meta) = format_data(data.copy())
         else:
-            if hasattr(data, "data"):
+            if isinstance(data, IamDataFrame):
                 # It might be a IamDataFrame?
                 (_df, _meta) = format_data(data.data.copy())
             else:
@@ -609,7 +618,7 @@ class ScmDataFrameBase(object):
         source_times_openscm = [
             convert_datetime_to_openscm_time(t) for t in self["time"]
         ]
-        if isinstance(target_times[0], datetime):
+        if isinstance(target_times[0], datetime.datetime):
             target_times_openscm = [
                 convert_datetime_to_openscm_time(t) for t in target_times
             ]
@@ -832,3 +841,160 @@ class ScmDataFrameBase(object):
         res["variable"] = res["variable"].apply(lambda x: "{} {}".format(x, append_str))
 
         return res.set_index(ts.index.names)
+
+    def append(self, other: ScmDataFrameBase, inplace=False, **kwargs):
+        """Appends additional timeseries from a castable object to the current dataframe
+
+        See ``df_append``
+
+        Parameters
+        ----------
+        other: openscm.scmdataframe.ScmDataFrame or something which can be cast to ScmDataFrameBase
+        """
+        if not isinstance(other, ScmDataFrameBase):
+            other = self.__class__(other, **kwargs)
+
+        return df_append([self, other], inplace=inplace)
+
+    def to_iamdataframe(self):
+        """Convert to  IamDataFrame instance
+
+        Returns
+        -------
+        An pyam.IamDataFrame instance containing the same data
+        """
+        if IamDataFrame is None:
+            raise ImportError(
+                "pyam is not installed. Features involving IamDataFrame are unavailable"
+            )
+
+        class LongIamDataFrame(IamDataFrame):
+            """This baseclass is a custom implementation of the IamDataFrame which handles datetime data which spans longer than pd.to_datetime
+            can handle
+            """
+
+            def _format_datetime_col(self):
+                if isinstance(self.data["time"].iloc[0], str):
+
+                    def convert_str_to_datetime(inp):
+                        return parser.parse(inp)
+
+                    self.data["time"] = self.data["time"].apply(convert_str_to_datetime)
+
+                not_datetime = [
+                    not isinstance(x, datetime.datetime) for x in self.data["time"]
+                ]
+                if any(not_datetime):
+                    bad_values = self.data[not_datetime]["time"]
+                    error_msg = "All time values must be convertible to datetime. The following values are not:\n{}".format(
+                        bad_values
+                    )
+                    raise ValueError(error_msg)
+
+        return LongIamDataFrame(self.timeseries())
+
+    def to_csv(self, path: str, **kwargs):
+        """Write timeseries data to a csv file
+
+        Parameters
+        ----------
+        path: string
+            file path
+        """
+        self.to_iamdataframe().to_csv(path, **kwargs)
+
+    def line_plot(self, x: str = "time", y: str = "value", **kwargs):
+        """Helper to generate line plots of timeseries
+
+        See ``pyam.IamDataFrame.line_plot`` for more information
+
+        """
+        return self.to_iamdataframe().line_plot(x, y, **kwargs)
+
+    def scatter(self, x: str, y: str, **kwargs):
+        """Plot a scatter chart using metadata columns
+
+        see pyam.plotting.scatter() for all available options
+        """
+        self.to_iamdataframe().scatter(x, y, **kwargs)
+
+    def region_plot(self, **kwargs):
+        """Plot regional data for a single model, scenario, variable, and year
+
+        see ``pyam.plotting.region_plot()`` for all available options
+        """
+        return self.to_iamdataframe().region_plot(**kwargs)
+
+    def pivot_table(self, index, columns, **kwargs):
+        """Returns a pivot table
+
+        see ``pyam.core.IamDataFrame.pivot_table()`` for all available options
+        """
+        return self.to_iamdataframe().pivot_table(index, columns, **kwargs)
+
+def df_append(dfs, inplace=False):
+    """
+    Append together many dataframes into a single ScmDataFrame
+    When appending many dataframes it may be more efficient to call this routine once with a list of ScmDataFrames, then using
+    `ScmDataFrame.append`. If timeseries with duplicate metadata are found, the timeseries are appended. For duplicate timeseries,
+    values fallings on the same timestep are averaged.
+    Parameters
+    ----------
+    dfs: list of ScmDataFrameBase object, string or pd.DataFrame.
+    The dataframes to append. Values will be attempted to be cast to non ScmDataFrameBase.
+    inplace : bool
+    If True, then the operation updates the first item in dfs
+    Returns
+    -------
+    ScmDataFrameBase-like object containing the merged data. The resultant class will be determined by the type of the first object
+    in dfs
+    """
+    dfs = [
+        df if isinstance(df, ScmDataFrameBase) else ScmDataFrameBase(df) for df in dfs
+    ]
+    joint_dfs = [df.copy() for df in dfs]
+    joint_meta = []
+    for df in joint_dfs:
+        joint_meta += df.meta.columns.tolist()
+
+    joint_meta = set(joint_meta)
+
+    # should probably solve this https://github.com/pandas-dev/pandas/issues/3729
+    na_fill_value = -999
+    for i, _ in enumerate(joint_dfs):
+        for col in joint_meta:
+            if col not in joint_dfs[i].meta:
+                joint_dfs[i].set_meta(na_fill_value, name=col)
+
+    # we want to put data into timeseries format and pass into format_ts instead of format_data
+    data = pd.concat(
+        [d.timeseries().reorder_levels(joint_meta) for d in joint_dfs], sort=False
+    )
+
+    data = data.reset_index()
+    data[list(joint_meta)] = data[joint_meta].replace(
+        to_replace=np.nan, value=na_fill_value
+    )
+    data = data.set_index(list(joint_meta))
+
+    data = data.groupby(data.index.names).mean()
+
+    if not inplace:
+        ret = dfs[0].copy()
+    else:
+        ret = dfs[0]
+
+    ret._data = data.reset_index(drop=True).T
+    ret._data = ret._data.sort_index()
+    ret["time"] = ret._data.index.values
+    ret._data = ret._data.astype(float)
+
+    ret._meta = (
+        data.index.to_frame()
+        .reset_index(drop=True)
+        .replace(to_replace=na_fill_value, value=np.nan)
+    )
+    ret._sort_meta_cols()
+
+    if not inplace:
+        return ret
